fix: :art: refactoring the lib directory structure
