--- src/interfaces/EmailProvider.ts
@@ -1,3 +1,0 @@
-export interface EmailProvider {
-  sendEmail(to: string, subject: string, body: string): Promise<void>;
-}
