+++ src/lib/index.ts
@@ -1,0 +1,3 @@
+export { EmailService } from "./services/EmailService";
+export { EmailProvider } from "./types";
+export { ProviderType } from "./types";
